==> 01_make_cst.Rmd <==
---
title: "CST and IST"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
refined = 'intermediates'
figure = 'working_figure'
private = 'data'
```

# DMM - Make CSTs and ISTs

Fits DMM models for rectal and nasal microbiota and tphe and ics immune data.  Selects optimal number of components and writes the relevant results.
```{r dmm}

#Start time
start.time <- Sys.time()

#Load packages
library(DirichletMultinomial)
library(lattice)
library(xtable)
library(parallel)

sites <- c("nas", "rec", "ics", "tphe")
maxClust <- 20 #Maximum number of clusters to try

for (site in sites) {

  #Read table with normalized, filtered, L7, counts of sample composition
  count <- t(as.matrix(read.csv(file.path(private, sprintf("%s_dmm_input.csv", site)), row.names=1)))
  
  
  #Fit the DMM with 1 through maxClust components
  fit <- mclapply(1:maxClust, dmn, count=count, verbose=TRUE, seed=100)
  
  #Get the Laplace measure of fit for each number of clusters
  lplc <- sapply(fit, laplace)
  
  #Determine the best number of clusters
  best_num_clusts <- which.min(lplc)
  
  #Pull out the fit for the best number of clusters
  best_model <- fit[[best_num_clusts]]
  
  #Write mixture weights, plot fit over number of clusters, and save the fitted model
  if(!(dir.exists(pth <- file.path(refined, 'dmn'))))
    dir.create(pth)
  write.csv(mixture(best_model), file = file.path(pth, sprintf("%s_mix_weights.csv", site)))
  plot(lplc, type="b", xlab="Number of Components", ylab="Model Fit")
  save(fit, file = file.path(pth, sprintf("%s_fit.rda", site)))
  
  #Print total runtime
  cat("Site runtime:")
  cat(difftime(Sys.time(), start.time, units = "hours"))
  cat("\n\n\n")

}

#Print total runtime
cat("Total runtime:")
cat(difftime(Sys.time(), start.time, units = "hours"))

```
