==> 04_tcell_associations.Rmd <==
---
title: "T cell / IST Associations"
author: "Andrew McDavid"
date: "`r Sys.Date()`"
output: github_document
---

```{r library, message = FALSE, warning = FALSE, results = 'hide', echo = FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
knitr::opts_chunk$set(cache = TRUE, autodep = TRUE)
knitr::opts_chunk$set(dev = c('png', 'pdf'))
library(broom)
library(tidyverse)
library(ggbeeswarm)
library(lme4)
library(car)

```

# ANOVA on some clusters

```{r}
all_feats = read_csv('intermediates/all_tcell_features.csv')
subject = read_csv('data/subject_covariates.csv')

# Define GA birth timepoints
cohort_groups = c(12, 5, -2)
# And bin samples by timepoint midpoints
mid_points = c(-10, 1.5, 8.5, 20)

subject = mutate(subject,  cohort = cut(preterm_weeks, breaks = mid_points, labels = c('-2', "5", "12")))


mc_suffix = tibble(Family = c('TPHE4', 'TPHE8', 'ICS4', 'ICS8'), 
                   suffix = c('t4', 't8', 'i4', 'i8'))
# Descriptive names for T cell subpop
metacluster_rn = read_csv('intermediates/Metacluster Identities.csv') %>% select(-X5) %>% mutate(Family = Family %>% toupper(), Family = str_replace_all(Family, ' ', '')) %>% left_join(mc_suffix) %>% mutate(marker = str_c('Meta.Cluster_', Cluster, '_', suffix))
```


```{r marker_anova}
marker_interest = gather(all_feats, key = 'marker', value = 'proportion', starts_with('Meta.Cluster')) %>% left_join(subject) %>% 
    mutate(cga_term = 2^cga - 37, 
           log2_cga_term_ratio = cga-log2(37)) %>%
    left_join(metacluster_rn %>% mutate(Identity = ifelse(is.na(Identity), marker, Identity) %>% str_replace_all('(?<![,])[ ]+', '') %>% abbreviate(minlength = 59),
                                        Identitywrap = str_wrap(Identity, 30)))

scale_cohort_color = function(...){
  scale_color_manual('Weeks premature', values = c('darkblue', 'purple', 'red'), ...)   
}


```



# Linear/non-linear metaclusters

```{r}
sin_sq = function(x) sin(x)^2
a_sqrt = function(x) asin(sqrt(x))

pred_values = expand.grid(preterm_weeks = cohort_groups, cga_term = seq(-12, 64, by = 4)) %>% filter(cga_term + preterm_weeks >= 0)

marker_nonlinear = marker_interest %>% group_by(marker, Identity, Category, Identitywrap) %>% do( output = 
    {
        model = lmer(a_sqrt(proportion) ~ (I( (cga_term<0)*cga_term ) + I( (cga_term>=0)*cga_term ))*preterm_weeks + (1|`Subject`), data = .)
        # Note I protects the indicator times cga_term, so these are not expanded.
        pre_contr = c(Intercept = 0, before = 1, after = 0, preterm = 0, prebefore = 10, preafter = 0)
        post_contr = c(Intercept = 0, before = 0, after = 1, preterm = 0, prebefore = 0, preafter = 10)
        pre10 = lht(model, pre_contr)[2,3]
        pre_est = fixef(model) %*% pre_contr
        post10 = lht(model, post_contr)[2,3]
        post_est = fixef(model) %*% post_contr
        diff_est = fixef(model) %*% (pre_contr - post_contr)
        diff10 = lht(model, pre_contr - post_contr)[2,3]
        monotone_tests = tibble(term = c('pre', 'post', 'diff'), estimate = c(pre_est, post_est, diff_est), p.value = c(pre10, post10, diff10))
        coefs = tidy(model, conf.int = TRUE)
        prediction = cbind(pred_values, .fitted = predict(model, newdata = pred_values, re.form = ~ 0 ))
        list(coefs = coefs, prediction = prediction, model = model, monotone_tests = monotone_tests)
             }) %>% rowwise() %>% mutate(model = list(output$model), coefs = list(output$coefs), prediction = list(output$prediction), monotone_tests = list(output$monotone_tests)) %>% select(-output)

nl_coefs = marker_nonlinear %>% unnest(coefs) %>% group_by(term) %>% mutate(p.value = pnorm(abs(statistic), lower.tail = FALSE)*2, p_adjust = p.adjust(p.value, method = 'bonferroni'))

nl_pred = marker_nonlinear %>% unnest(prediction)

traj_plot = ggplot(nl_pred %>% filter(marker == 'Meta.Cluster_10_i4'), aes(x = cga_term)) + geom_line(aes(color = factor(preterm_weeks), y = .fitted), lwd = 2) + scale_cohort_color() #+ geom_ribbon(aes(ymin = .fitted - .se.fit, ymax = .fitted + .se.fit), alpha = .3)

```

## Prematurity associated (Figure 2D)

```{r preterm_summary, fig.width = 5, fig.height = 7}
preterm_markers = nl_coefs %>% filter(term == 'preterm_weeks', p_adjust < .05) %>% mutate(Identity = fct_reorder(factor(Identity), -estimate), rank = rank(p_adjust))

ggplot(preterm_markers, aes(x = Identity, y = -estimate, ymin = -conf.low, ymax = -conf.high)) + geom_pointrange(aes(color = Category)) + ylab('Change in population at 37 week PMA \n per week increase of gestational age at birth') + geom_hline(yintercept = 0, lty = 2)  + geom_text(aes(label = Identity, y = -estimate+.002), size = 2, hjust = 0) + theme(legend.position = 'none') +  coord_flip() + facet_grid(Category ~ ., scales = 'free_y', space = 'free_y') + scale_x_discrete(labels = NULL, breaks = NULL)
```


## Figure 2E

```{r preterm_markers}
other_preterm_toplot = str_c('Meta.Cluster_', c('26_t4', '25_t4', '1_t8', '2_t8', '18_i4', '13_i4', '12_i8', '8_i8'))

traj_plot %+% (nl_pred %>% semi_join(filter(preterm_markers, rank < 6), by = 'marker')) + facet_wrap(~Identitywrap, scales = 'free_y')  + geom_jitter(data = marker_interest %>% semi_join(filter(preterm_markers, rank < 6)), aes(y = a_sqrt(proportion), color = cohort), size = 1, alpha = .4) + theme(strip.text = element_text(size = 8))

traj_plot %+% (nl_pred %>% semi_join(filter(preterm_markers, marker %in% other_preterm_toplot), by = 'marker')) + facet_wrap(~Identitywrap, scales = 'free_y')  + geom_jitter(data = filter(marker_interest, marker %in% other_preterm_toplot), aes(y = a_sqrt(proportion), color = cohort), size = 1, alpha = .4) + theme(strip.text = element_text(size = 8))

```



Trajectories for top 6 markers with GAB differences at 37 weeks PMA.

***

## Non-monotone markers (Figure S1)

```{r}
nl_monotone = unnest(marker_nonlinear, monotone_tests)
nonmono_candidates = nl_monotone %>% group_by(marker) %>% summarize(max_sig = max(p.value), n_sig = sum(p.value < .05))
nonmono_sign = nl_monotone %>% select(-p.value) %>% spread(term, estimate) %>% mutate(signchange = sign(post) != sign(pre))

nonmono_candidates = left_join(nonmono_candidates, nonmono_sign) %>% filter(signchange, n_sig ==3, p.adjust(max_sig/3, method = 'bonferroni')< .05) %>% mutate(rank = rank(max_sig)) %>% left_join(nl_monotone %>% select(-Identity:-Identitywrap), by = c('marker'))  %>% mutate(Identitywrap = fct_reorder(factor(Identitywrap), rank))

ggplot(nonmono_candidates  %>% filter(term != 'diff'), aes(x = Identitywrap, y = estimate, color = term)) + geom_point() + ylab('Trajectory difference at 37 week PMA for 25 week CGA') + geom_hline(yintercept = 0, lty = 2) +  coord_flip() 
```

These markers have an inversion estimated for 25 week-GAB babies. 

```{r non_mono_all, fig.width = 8, fig.height = 6}
traj_plot %+% (nl_pred %>% semi_join(filter(nonmono_candidates, rank < Inf), by = 'marker')) + facet_wrap(~Identitywrap, scales = 'free_y')+ geom_jitter(data = marker_interest %>% semi_join(filter(nonmono_candidates, rank < Inf)), aes(y = a_sqrt(proportion), color = cohort), size = 1, alpha = .4) + theme(strip.text = element_text(size = 5))  + aes(x = cga_term + 37) + xlab("PMA (weeks)") + ylab("asin(sqrt(p))")
```

Estimated trajectories and observed data for top 6 non-monotone markers.

```{r trajectories_cleanup, fig.height = 4, fig.width = 8}
fig5_marker_toplot = tibble(marker = c('Meta.Cluster_13_i4', 'Meta.Cluster_18_i4', 'Meta.Cluster_8_i8', 'Meta.Cluster_12_i8',
                                  'Meta.Cluster_26_t4', 'Meta.Cluster_25_t4', 'Meta.Cluster_1_t8', 'Meta.Cluster_2_t8'))

strReverse = function(x) sapply(lapply(strsplit(x, NULL), rev), paste, 
collapse="")

#nl_fig5 = semi_join(nl_pred, fig5_marker_toplot) %>% mutate(Identity = factor(Identity, levels = sort(unique(strReverse(marker)))))

(traj_plot %+% semi_join(nl_pred, fig5_marker_toplot)) + facet_wrap(~Identitywrap, scales = 'free', nrow = 2)+ geom_jitter(data = semi_join(marker_interest, fig5_marker_toplot), aes(y = a_sqrt(proportion), color = cohort), size = 1, alpha = .4) + theme(strip.text = element_text(size = 5))  + aes(x = cga_term + 37) + xlab("PMA (weeks)") + ylab("asin(sqrt(p))")

```

# Immune samples/timepoints

```{r, results = 'asis'}
flow_subj_counts = all_feats %>% select(Subject, `Sequence Num`, tphe = Meta.Cluster_11_t8, ics=Meta.Cluster_11_i8) %>% mutate(both_tphe_ics = !is.na(ics) & !is.na(tphe))

knitr::kable(flow_subj_counts %>% group_by(`Sequence Num`, both_tphe_ics) %>% summarize(n()))
```

# IST proportions

```{r}
# IST posterior probs per subject/time point in long format
all_ist = read_csv('data/fitted_ist_longformat.csv') %>% rename(Subject = Alias)

```


```{r ist_by_term}
all_ist_cat = all_ist %>% group_by(Subject, `Sequence Num`, assay) %>% summarize(IST_prob = max(value), IST_modal = IST[which.max(value)])

all_ist_cat = left_join(all_ist_cat, subject, by = "Subject") %>% mutate(type = assay) %>% group_by(type, cohort) %>% mutate(total_obs = length(type))

tests_by_ist = map_dfr(unique(all_ist_cat$IST_modal), ~ {
    type = str_extract(.x, 'ICS|TPHE')
    dat = filter(all_ist_cat, type == type)
    model = glm(IST_modal ==  .x ~ preterm_weeks, family = 'binomial', data = dat)
    tidy(model, conf.int = TRUE) %>% mutate(ist = .x)
})

tests_by_ist = tests_by_ist %>% group_by(term) %>% mutate(p_adjust = p.adjust(p.value, method = 'bonferroni'))

ggplot(filter(tests_by_ist, term == 'preterm_weeks'), aes(x = ist, y = estimate, ymin = conf.low, ymax = conf.high, shape = p_adjust < .05)) + geom_pointrange() + ylab('Change in log odds per week pre-term') + geom_hline(yintercept = 0, lty = 2) +  coord_flip() + scale_shape_discrete('Bonferroni p < .05')

```


# IST-CMV-Antibiotic associations

### Setup covariates

```{r}
# cmv+ : positive for 2 targets via PCR on saliva at one year
# lab+ : positive for serology on blood at one year
cmv_dat = subject %>% mutate(cmv_pcr = str_detect(`cmv test`, fixed('cmv+')))

hospital_humilk = read_csv('data/milk_hospital.csv') %>% rename(perinatal_milk = `Any Human Milk Perinatal`)
nabx_polish = read_csv('data/antibiotic_exposure.csv')
nabx_time = nabx_polish %>% select(-group) %>% rename(n_antibiotics = `Number of systemic antibiotic`) %>% spread(discharge, n_antibiotics) %>% rename(n_antibiotics_discharge = 'TRUE', n_antibiotics_pre = 'FALSE')

discharge_humilk = read_csv( 'intermediates/milk_subject.csv')

covariates =  purrr::reduce(list(cmv_dat, hospital_humilk, nabx_time, discharge_humilk), left_join) %>%
  mutate(Race = fct_collapse(Race, other_unk = c('Other Race', 'Unknown or Not Reported', 'Asian')),
         preg_membrane_18hr  = fct_collapse(preg_membrane_18hr, No_Unknown = c('No', 'Unknown')))
covariates$n_antibiotics_pre = ifelse(covariates$preterm_weeks <=0, NA, covariates$n_antibiotics_pre)
```

### Generate presence/absence tables of ists per early/late/ever epoch

```{r}
get_ist_tables = function(x, label){
  ist_tab = x %>% ungroup(all_ist_cat) %>% 
  group_by(assay) %>% do({
    sub = factor(.$Subject)
  tab = table(Var1 = sub, .$IST_modal)
  as.data.frame(tab) %>% 
    dplyr::rename(Subject = Var1, IST = Var2) %>%
    mutate(present = Freq > 0, epoch = label)
}) %>% left_join(covariates)
}
 
ist_ever = filter(all_ist_cat, `Sequence Num` %in% c(1,7, 19)) %>% get_ist_tables(label = 'ever')
ist_late = filter(all_ist_cat, `Sequence Num` == 19) %>% get_ist_tables(label = 'early')
ist_early = filter(all_ist_cat, `Sequence Num` < 19)  %>% get_ist_tables(label = 'late')
```


### Antibiotics, CMV, Inflammation and TPHE5 (Figure S2)

```{r}
library(mgcv)
combined_epochs = bind_rows(ist_ever, ist_late, ist_early)

cleanup_gam = function(gamfit){
  gamfit_ci = tidy(gamfit, parametric = TRUE) %>% mutate(conf.low = estimate - 1.96*std.error, conf.high = estimate + 1.96*std.error )
  fit_sub = filter(gamfit_ci, term != '(Intercept)', !str_detect(term, 'Race')) %>% mutate(p_adjust = p.adjust(p.value, 'bonferroni'), term = fct_reorder(term, p.value), pval = as.character(cut(p.value, breaks = c(0, .01, .05, 1), labels = c('**', '*', ''))), n_obs = nobs(gamfit))
  fit_sub
}


gamss = combined_epochs %>% group_by(IST, epoch) %>% do({
  stopifnot(nrow(.) <= 267)
  model_cmv_ = gam(present ~ s(preterm_weeks) + preg_antibiotics + cchorio + mode_delivery + preg_membrane_18hr + Race + Gender + cmv_pcr, data = ., family = 'binomial')
  model_milk_ = gam(present ~  s(preterm_weeks) +  perinatal_milk  + milk_months + mode_delivery + Race + Gender, data = ., family = 'binomial')
model_abx_pre_ = gam(present ~  s(preterm_weeks) + n_antibiotics_pre  + mode_delivery + Race + Gender, data = ., family = 'binomial')
  model_abx_post_ = gam(present ~  s(preterm_weeks) + n_antibiotics_discharge + mode_delivery + Race + Gender, data = ., family = 'binomial')
  tibble(cmv_df = list(cleanup_gam(model_cmv_)), model_cmv = list(model_cmv_), 
         model_abx_pre = list(model_abx_pre_),
         model_abx_post = list(model_abx_post_),
         model_milk= list(model_milk_),
         milk_df = list(cleanup_gam(model_milk_)),
         abx_pre_df = list(cleanup_gam(model_abx_pre_)),
         abx_post_df = list(cleanup_gam(model_abx_post_)))
         
})

```


```{r}
tphe_tests_sub = filter(gamss, IST %in% c('TPHE_5', 'TPHE_6'), epoch == 'ever') %>% ungroup() %>% select(IST, cmv_df) %>% unnest(cols = c(cmv_df))


tphe5_plt = ggplot(tphe_tests_sub, aes(x = term, y = estimate, ymin = conf.low, ymax = conf.high)) + geom_pointrange(position = position_dodge(width =.3)) + geom_text(aes(label = pval), size = 7, position = position_nudge(x = -.1, y = -.2)) + ylab('Log odds of ever TPHE5') + geom_hline(yintercept = 0, lty = 2) + theme(axis.text.x = element_text(angle = 90))

```


```{r tphe5_6_assoc, dev = c('png', 'pdf'), fig.width = 7, fig.height = 4}
tphe5_plt + aes(color = IST) + ylab('Log odds\nIST ever') 

```

parametric effects

## TPHE5 vs TPHE6 associations

```{r}
early_late = filter(combined_epochs, epoch %in% c('early', 'late'), IST %in% c('TPHE_5', 'TPHE_6')) %>% ungroup()

early_late_crosstab = pivot_wider(early_late %>% select(Subject, IST, present, epoch), names_from = c(IST, epoch), values_from = present)

with(early_late_crosstab, fisher.test(TPHE_5_early, TPHE_6_late))
with(early_late_crosstab, fisher.test(TPHE_6_early, TPHE_5_late))
```


##  Other inflam associations

```{r}
all_assoc = gamss %>% ungroup() %>% select(IST, epoch, cmv_df) %>% unnest(cols = c(cmv_df)) %>%  group_by(term) %>% mutate(termwise_fdr = p.adjust(p.value)) %>% arrange(p.value) %>% select(-p_adjust, -pval)

knitr::kable(head(all_assoc, n= 10), digits = 3)
write_csv(all_assoc, 'intermediates/inflam_ics_assocs.csv')
```

Top 20 associations, full table under ['intermediates/inflam_ics_assocs.csv'](intermediates/inflam_ics_assocs.csv)

## Milk associations

```{r}
milk_assoc = gamss %>% ungroup() %>% select(IST, epoch, milk_df) %>% unnest(cols = c(milk_df)) %>%  group_by(term) %>% mutate(termwise_fdr = p.adjust(p.value)) %>% arrange(p.value) %>% select(-p_adjust, -pval)

knitr::kable(head(milk_assoc, n= 10), digits = 3)
write_csv(milk_assoc, 'intermediates/milk_ics_assocs.csv')
```

Top 20 associations, full table under ['intermediates/milk_ics_assocs.csv'](intermediates/milk_ics_assocs.csv)


## Antibiotic associations (hospital)

```{r}
abx_assoc = gamss %>% ungroup() %>% select(IST, epoch, abx_pre_df) %>% unnest(cols = c(abx_pre_df)) %>%  group_by(term) %>% mutate(termwise_fdr = p.adjust(p.value)) %>% arrange(p.value) %>% select(-p_adjust, -pval)

knitr::kable(head(abx_assoc, n= 10), digits = 3)
write_csv(abx_assoc, 'intermediates/abx_pre_df_ics_assocs.csv')
```

Top 20 associations, full table under ['intermediates/abx_pre_df_ics_assocs.csv'](intermediates/abx_pre_df_ics_assocs.csv)

## Antibiotic associations (discharge)

```{r}
abx_post_assoc = gamss %>% ungroup() %>% select(IST, epoch, abx_post_df) %>% unnest(cols = c(abx_post_df)) %>%  group_by(term) %>% mutate(termwise_fdr = p.adjust(p.value)) %>% arrange(p.value) %>% select(-p_adjust, -pval)

knitr::kable(head(abx_post_assoc, n = 10), digits = 3)
write_csv(abx_post_assoc, 'intermediates/abx_post_df_ics_assocs.csv')
```

Top 20 associations, full table under ['intermediates/abx_post_df_ics_assocs.csv'](intermediates/abx_post_df_ics_assocs.csv)
